feat: 6487 restructure Python packages
