--- pysparkle/pyproject.toml
@@ -1,31 +1,0 @@
-[tool.poetry]
-name = "pysparkle"
-version = "0.1.4"
-description = "PySpark data transformer"
-authors = ["ensono.com"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.9"
-click = "^8.1.3"
-pyspark = "^3.4.0"
-azure-identity = "^1.13.0"
-azure-storage-file-datalake = "^12.11.0"
-click-loglevel = "^0.4.0.post1"
-colorlog = "^6.7.0"
-great-expectations = "^0.16.0"
-pydantic = "^1.10.9"
-
-[tool.poetry.scripts]
-pysparkle = "pysparkle.pysparkle_cli:cli"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^7.3.1"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.pytest.ini_options]
-log_cli = true
-log_cli_level = "INFO"
